Created category helpers and services for delete
