+++ pages/api/controllers/categoryControllers.js
@@ -1,0 +1,5 @@
+import { deleteServices } from '../services/categoryServices';
+
+export const deleteController = async (body, res) => {
+    deleteServices(body, res);
+};
